+++ src/uts.rs
@@ -27,15 +27,15 @@ pub struct Uts {
 
     /// Hours associated with the sound.
     #[serde(skip_serializing_if = "Option::is_none", rename = "Hours")]
-    pub hours: Option<NwValue<u8>>,
+    pub hours: Option<NwValue<u32>>,
 
     /// Interval of the sound.
     #[serde(skip_serializing_if = "Option::is_none", rename = "Interval")]
-    pub interval: Option<NwValue<i32>>,
+    pub interval: Option<NwValue<u32>>,
 
     /// Interval variation of the sound.
     #[serde(skip_serializing_if = "Option::is_none", rename = "IntervalVrtn")]
-    pub interval_vrtn: Option<NwValue<i32>>,
+    pub interval_vrtn: Option<NwValue<u32>>,
 
     /// Localized name of the sound.
     #[serde(skip_serializing_if = "Option::is_none", rename = "LocName")]
